pyproject.toml
[project]
name = "pymailai"
version = "0.1.0"
description = "Email interface for AI agents"
authors = ["Ivan Iufriakov"]
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "python-dotenv>=1.0.0",
    "aiosmtplib>=2.0.2",
    "aioimaplib>=1.0.1",
    "email-validator>=2.0.0",
    "google-auth>=2.22.0",
    "google-auth-oauthlib>=1.0.0",
    "google-api-python-client>=2.95.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-mock>=3.10.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "isort>=5.0.0",
    "flake8>=6.1.0",
    "mypy>=1.0.0",
]

openai = [
    "openai>=1.0.0",
    "pydantic>=2.0.0",
]

anthropic = [
    "anthropic>=0.3.0",
]

all = [
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "anthropic>=0.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.black]
line-length = 88
target-version = ["py38"]

[tool.isort]
profile = "black"
multi_line_output = 3
